refactor: add src folder; change lib to services
