--- cmd/global-dispatcher/dispatch-globally/cli/main.go
@@ -1,34 +1,0 @@
-package main
-
-import (
-	"context"
-	"time"
-
-	base "github.com/Pocket/global-services/cmd/global-dispatcher/dispatch-globally"
-	"github.com/Pocket/global-services/common/environment"
-
-	logger "github.com/Pocket/global-services/lib/logger"
-	"github.com/Pocket/global-services/lib/utils"
-	log "github.com/sirupsen/logrus"
-)
-
-var timeout = time.Duration(environment.GetInt64("TIMEOUT", 360)) * time.Second
-
-func main() {
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-
-	requestID, _ := utils.RandomHex(32)
-	failedDispatcherCalls, err := base.DispatchSessions(ctx, requestID)
-	if err != nil {
-		logger.Log.WithFields(log.Fields{
-			"requestID":      requestID,
-			"error":          err.Error(),
-			"failedDispatch": failedDispatcherCalls,
-		}).Error("ERROR DISPATCHING SESSION: " + err.Error())
-	}
-	logger.Log.WithFields(log.Fields{
-		"requestID":      requestID,
-		"failedDispatch": failedDispatcherCalls,
-	}).Info("GLOBAL DISPATCHER RESULT")
-}
