+++ src/utils/sliceToMap.go
